// Frontend/src/types/api.ts
export interface PhoneNumber {
  number: string;
  prefers_texting?: boolean;
  no_english?: boolean;
}

export interface LineItem {
  subject: string;
  price_cents: number;
}

export interface ClockLog {
  user: string;
  out: boolean;
  timestamp: number;
}

// TinyTicket interfaces for lists
export interface TinyTicket {
  ticket_number: number;
  subject: string;
  customer_id: string;
  status: string;
  device: string;
  created_at: number;
  customer_name: string;
}

export interface TicketWithoutCustomer {
  ticket_number: number;
  subject: string;
  status: string;
  password?: string;
  device: string;
  created_at: number;
  last_updated: number;
  comments?: Comment[];
  attachments?: string[];
  items_left?: string[];
  line_items?: LineItem[];
  paid_at?: number;
  total_paid_cents?: number;
  customer_name?: string;
}

export interface Ticket extends TicketWithoutCustomer {
  customer: Customer;
}

export interface Customer {
  customer_id: string;
  full_name: string;
  email?: string;
  phone_numbers: PhoneNumber[];
  created_at: number; // Backend is i64
  last_updated: number; // Backend is i64
}

export interface Comment {
  comment_body: string;
  tech_name: string;
  created_at: number; // Backend is i64
}

export interface CognitoUser {
  username: string;
  email: string | null;
  given_name: string | null;
  enabled: boolean;
  groups: string[];
  created: string | null;
  user_status: string;
}

export interface PostTicket {
  customer_id: string;
  subject: string;
  password: string | null;
  items_left: string[] | null;
  device: string;
}

export interface UpdateTicket {
  subject: string | null;
  password: string | null;
  items_left: string[] | null;
  line_items: LineItem[] | null;
  device: string | null;
}

export interface PostCustomer {
  full_name: string;
  email: string | null;
  phone_numbers: PhoneNumber[];
}

export interface UpdateCustomer {
  full_name: string | null;
  email: string | null;
  phone_numbers: PhoneNumber[] | null;
}

export interface PostComment {
  comment_body: string;
}

export interface PostInviteUser {
  email: string;
  firstName: string;
}

export interface PostUpdateUserGroup {
  username: string;
  group: string;
}

export interface PostAttachment {
  image_data: string;
}

export interface StoreConfig {
  store_name: string;
  tax_rate: number;
  address: string;
  city: string;
  state: string;
  zip: string;
  phone: string;
  email: string;
  disclaimer: string;
}